requirements.txt
# Development dependencies
ruff
pre-commit

# Add your project dependencies below
requests
python-dotenv
Flask